requirements.txt
mock
pytest
PyYAML==3.10
ordereddict
Django>=1.5
pytest-django
Flask
CherryPy
selenium
pytest-cov
argparse
Jinja2>=2.7
six
lettuce
http://github.com/pivotal/jasmine/tarball/python_2#egg=jasmine-core
http://github.com/mockfs/mockfs/tarball/master#egg=mockfs
